utils: Fix write_u32 prototype

+++ src/utils.h
@@ -58,7 +58,7 @@ static inline uint8_t *write_u16(uint8_t *p, const uint16_t v)
         return (p);
 }
 
-static inline uint8_t *write_u32(uint8_t *p, const uint16_t v)
+static inline uint8_t *write_u32(uint8_t *p, const uint32_t v)
 {
         *p++ = (v >> 24) & 0xFF;
         *p++ = (v >> 16) & 0xFF;
